Move generic parts of intrinsic interface out of eos-vm-oc.hpp
