+++ libraries/rodeos/rodeos.cpp
@@ -242,8 +242,8 @@ std::once_flag registered_filter_callbacks;
 rodeos_filter::rodeos_filter(eosio::name name, const std::string& wasm_filename
 #ifdef EOSIO_EOS_VM_OC_RUNTIME_ENABLED
                              ,
-                             const boost::filesystem::path&                    eosvmoc_path,
-                             const eosio::chain::webassembly::eosvmoc::config& eosvmoc_config, bool eosvmoc_enable
+                             const boost::filesystem::path&       eosvmoc_path,
+                             const eosio::chain::eosvmoc::config& eosvmoc_config, bool eosvmoc_enable
 #endif
                              )
     : name{ name } {
